feat(jobs): Update

+++ src/utils.rs
@@ -1,10 +1,10 @@
 use std::time::{SystemTime, UNIX_EPOCH};
 
-pub fn get_current_timestamp() -> u64 {
+pub fn get_current_timestamp() -> i32 {
     let start = SystemTime::now();
     let since_the_epoch = start
         .duration_since(UNIX_EPOCH)
         .expect("Time went backwards");
 
-    since_the_epoch.as_secs()
+    since_the_epoch.as_secs() as i32
 }
